WIP: Refactor menu
