+++ src/funcDeclarations.hpp
@@ -2,6 +2,8 @@
 // Created by mayur on 14/10/22.
 // Function declarations
 
+/*
 int welcome(Input);
 int validation(Input);
 int userInput();
+*/
